storage: improved stats performance. (skipping sample names query)
Previously, each time a variant was fetched from the DBAdaptor
(DBIterator actually), the inner DBObjectToSamplesConverter made
another query to Mongo to find out the sample names.

In main.statsVariants(), this is a waste because every variant will
share sourceEntry, so the names list will allways be the same. For that,
the adaptor now can be told to fetch the sample names just once at the
beginning.

Conflicts cherry-picking (took ours):
	opencga-storage/opencga-storage-core/src/main/java/org/opencb/opencga/storage/core/runner/StringDataWriter.java
	opencga-storage/opencga-storage-core/src/main/java/org/opencb/opencga/storage/core/variant/io/VariantDBReader.java
	opencga-storage/opencga-storage-core/src/main/java/org/opencb/opencga/storage/core/variant/stats/VariantStatisticsManager.java