storage: Problem solved when saving the StudyConfiguration after closing the mongo conexion.
In the Load step, the VariantMongoDBWriter, at the end, closes the mongoDB conection.
DataStore uses a pool of conections, so, if you close one conexion, the conection is closed for everybody.
If the conection for saving the StudyConfiguration and the variants (by the VariantMongoDBWriter) was
the same (same DB) the MongoDBStudyConfigurationManager was unable to update the StudyConfiguration.

Now, every time that the StudyConfigurationManager is obtained, gets a new conection. If the conection was alive,
it is taken from the pool of conections, else, creates a new one.