storage: VariantMongoDBAdaptor insert improvements for general case
When pushing a new study (when the study does not exists in this variant, or the variant does not exist),
instead of inserting an empty study, include also the file and genotype information.
Then, push the file and genotype information to those variants that already exist and had the study.