storage:vcf dump: improved support for multiallelic, phased, and indels
regarding the indel support:
If an attribute "ori":"AT	A" is present, that alleles will be used.
Otherwise, the variant will not be written because an empty allele is not
VCF-compliant.

multiallelic: variants will be written in separate variants, one for each
alternate allele. The secondary alleles will appear as ".", as the allele
doesn't appear in that variant.