storage: fixed bug with VariantSourceMongoDBAdaptor sample cache.
Before, when the samples of a fileId were asked, the condition to retrieve them
from the "files" collection was that the number of files wasn't the same. This
is wrong.

The fix is: when the samples of a fileId are asked, look up those fileIds in
the cache, and if there is anyone missing, empty the cache, and dump the
samplenames from the "files" collection.