storage: Clean stage collection while loading data.
This will avoid a collection-scan over the STAGE collection to find the
documents to clean. Also it will allow to make a more acurate cleaning,
knowing if the study or the files must be removed, avoiding two
collection-scan.

It also will speed up the merge step in case of resume. Will skip those
variants already cleaned (i.e. already processed)