Several identifier and job related fixes: - JobMongoDBAdaptor supports search by ID, as JobManager assumes - ToolAnalysis.execute search the job by UID (what it knows) - also it looks for the study FQN, what it needed instead of the study UID - ToolAnalysisTest queries by projectStr instead of projectUID
There are still errors that are not easy to fix without changing the interfaces and settling on how resources should be searched for. For example, ResourceManager.get(long jobId, QueryOptions options, String sessionId) received an UID, but then

public QueryResult<Job> get(long jobId, QueryOptions options, String sessionId) throws CatalogException {
    return get(null, String.valueOf(jobId), options, sessionId);
}

casts it to a String and assumes that it is an FQN (an ID). This cannot work.